feat(ui): expose GRAPH_SETTING_DEFAULTS from store-free utils entry
Move GRAPH_SETTING_DEFAULTS into its own dependency-free module
(config/graphSettingDefaults.ts) and re-export it (plus a
GraphSettingDefaults type) from components/utils.ts so prop-driven
consumers of PixiGraphCanvas + PhysicsPanel can import the defaults
from @opentrace/opentrace/utils without pulling in the store,
providers, or pixi.

useGraphViewer now imports the constant from the new module, keeping
it the single source of truth. Values are unchanged.
